完成员工禁用功能;已知晓的bug:员工分页查询总数为NaN

+++ src/main/java/com/reaosen/management_system/DTO/UserDTO.java
@@ -5,9 +5,9 @@ import lombok.Data;
 
 @Data
 public class UserDTO {
-    private Long id;
+    private Integer id;
     @JSONField(name = "id")
-    private Long accountId;
+    private Integer accountId;
     private String email;
     private String username;
     private String avatarUrl;
